Slightly lazy fix for Issue #1 - "Sometimes a shortcut is possible but not marked as correct"
We just see if there is more than one path through the maze, if there is then we regenerate the maze (up to a limit of attempts so as to prevent inf loop).

A better solution is to upfront pick a start and end point, then generate the path through the maze avoiding the 8 cells around the finish (unless its at an edge/corner) til we get to the last point. This will invovle implementing backtracking etc.

Anyway this is good enough